Introduce quickcheck testing
This commit introduces the quickcheck library into the project,
testing the word/is_* traits. Quickcheck is very helpful for finding
edge-case bugs in complex algorithms -- of which the is_* are not --
so this commit exists as more of a demonstration of technique. Macros
are used to generate test code over built-in integer types and slower
but "obviously correct" algorithms are tested against the production
quality code.

I'd like to use bitwise in a compact datastructure project which I
expect to QC and fuzz like crazy. To that end I thought it wise to
introduce QC'ing directly into bitwise and, if you're open to it,
fuzzing as well via [cargo-fuzz](https://github.com/rust-fuzz/cargo-fuzz)
in a latter commit. QC testing is less efficient at detecting bugs
compared to exhaustive testing where exhaustive testing is applicable
but _pretty good_ at searching the haystack in large domains like u64.
This can be further improved with branch-directed fuzzing.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>